conflict fleetManager
