+++ controller/fleetController.go
@@ -43,7 +43,7 @@ func AssignShipToFleet(w http.ResponseWriter, r *http.Request) {
 		panic(exception.NewHttpException(400, "Wrong location", nil));
 	}
 	
-	manager.AssignShipToFleet(ship,fleet)
+	shipManager.AssignShipToFleet(ship,fleet)
 	
     utils.SendJsonResponse(w, 200,fleet)
 }
@@ -75,7 +75,7 @@ func RemoveShipFromFleet(w http.ResponseWriter, r *http.Request){
 	if ship.Fleet == nil{ // normaly this does not triger if the data of ship.IsShipInFleet is correct
 		panic(exception.NewHttpException(400, "Ship already is not in a fleet", nil));
 	}
-	manager.AssignShipToHangard(ship)
+	shipManager.AssignShipToHangard(ship)
 	
     utils.SendJsonResponse(w, 200, nil /*TODO*/) // What do I return ?
 }
